refactor(voicecli): move nats/ into adapters/nats/ (T1 of #83)
Wholesale dir move + 35 import-site rewrites + pyrightconfig + .importlinter
patches. No behavior change.
