+++ .importlinter
@@ -9,7 +9,7 @@ include_external_packages = False
 name = Layer dependency rule
 type = layers
 layers =
-    voicecli.cli | voicecli.nats | voicecli.tg | voicecli.listen | voicecli.overlay
+    voicecli.cli | voicecli.adapters.nats | voicecli.tg | voicecli.listen | voicecli.overlay
     voicecli.api
     voicecli.engine | voicecli.transcribe | voicecli.daemon | voicecli.stt_daemon | voicecli.model_registry
     voicecli.config | voicecli.paths | voicecli.utils | voicecli.env | voicecli.models | voicecli.samples | voicecli.translate | voicecli.markdown | voicecli.stt_modes
@@ -18,10 +18,10 @@ ignore_imports =
     voicecli.translate -> voicecli.markdown
     # Transitional: voicecli.translate imports voicecli.utils for shared utilities — same layer peer dependency
     voicecli.translate -> voicecli.utils
-    # Transitional: voicecli.cli imports voicecli.nats adapters inline for NATS subcommands — needs extraction to adapter layer
-    voicecli.cli -> voicecli.nats.tts_adapter
-    voicecli.cli -> voicecli.nats.config
-    voicecli.cli -> voicecli.nats.stt_adapter
+    # Transitional: voicecli.cli imports voicecli.adapters.nats adapters inline for NATS subcommands — needs extraction to adapter layer
+    voicecli.cli -> voicecli.adapters.nats.tts_adapter
+    voicecli.cli -> voicecli.adapters.nats.config
+    voicecli.cli -> voicecli.adapters.nats.stt_adapter
     # Transitional: voicecli.cli imports voicecli.listen for recording subcommand — needs extraction to adapter layer
     voicecli.cli -> voicecli.listen
     # Transitional: voicecli.daemon imports voicecli.engine to load TTS engines — daemon is an infrastructure adapter, coupling is expected but needs SynthesisPort wiring (ADR-059)
