+++ src/voicecli/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""VoiceCLI adapters layer — port implementations."""
