+++ src/voicecli/adapters/nats/_validate.py
