+++ src/voicecli/adapters/nats/stt_adapter.py
@@ -14,8 +14,8 @@ from typing import Any
 from roxabi_contracts.envelope import CONTRACT_VERSION
 from roxabi_contracts.voice.models import SttResponse
 from roxabi_nats import NatsAdapterBase
-from voicecli.nats.queue_groups import STT_WORKERS
-from voicecli.nats.tempdir import cleanup, scoped_path
+from voicecli.adapters.nats.queue_groups import STT_WORKERS
+from voicecli.adapters.nats.tempdir import cleanup, scoped_path
 
 # voicecli.api is NOT imported at module level — deferred to keep startup fast
 # and avoid pulling torch/faster-whisper when only inspecting the adapter (e.g. --help).
