+++ tests/nats/test_tts_adapter.py
@@ -22,9 +22,9 @@ import pytest
 # ---------------------------------------------------------------------------
 
 try:
-    from voicecli.nats.config import _resolve_engine
-    from voicecli.nats.tempdir import scoped_path  # noqa: F401
-    from voicecli.nats.tts_adapter import TtsNatsAdapter
+    from voicecli.adapters.nats.config import _resolve_engine
+    from voicecli.adapters.nats.tempdir import scoped_path  # noqa: F401
+    from voicecli.adapters.nats.tts_adapter import TtsNatsAdapter
 
     _IMPORT_ERROR: ImportError | None = None
 except ImportError as _e:
@@ -140,7 +140,9 @@ class TestTtsNatsAdapter:
         with patch(
             "voicecli.engine._get_registry", return_value={"mock": _stub_engine_factory(tmp_path)}
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         # Assert
@@ -171,7 +173,9 @@ class TestTtsNatsAdapter:
         with patch(
             "voicecli.engine._get_registry", return_value={"mock": _stub_engine_factory(tmp_path)}
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         reply = msg.last_reply()
@@ -211,7 +215,9 @@ class TestTtsNatsAdapter:
             "voicecli.engine._get_registry",
             return_value={"mock": _stub_engine_factory(tmp_path, raises=RuntimeError("model OOM"))},
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         # Assert
@@ -233,7 +239,9 @@ class TestTtsNatsAdapter:
         with patch(
             "voicecli.engine._get_registry", return_value={"mock": _stub_engine_factory(tmp_path)}
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         # Assert — succeeds normally, reply stamps contract_version "1"
@@ -348,10 +356,12 @@ class TestTtsNatsAdapter:
             return real_b64encode(buf)
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
-            patch("voicecli.nats.tts_adapter.base64.b64encode", side_effect=_sniff),
+            patch("voicecli.adapters.nats.tts_adapter.base64.b64encode", side_effect=_sniff),
         ):
             asyncio.run(adapter.handle(msg, payload))
 
@@ -378,7 +388,9 @@ class TestTtsNatsAdapter:
         with patch(
             "voicecli.engine._get_registry", return_value={"mock": _stub_engine_factory(tmp_path)}
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         # Assert — temp file removed after successful reply
@@ -403,7 +415,9 @@ class TestTtsNatsAdapter:
             "voicecli.engine._get_registry",
             return_value={"mock": _stub_engine_factory(tmp_path, raises=RuntimeError("kaboom"))},
         ):
-            with patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path):
+            with patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ):
                 asyncio.run(adapter.handle(msg, payload))
 
         # Assert — temp file removed even when engine raises
@@ -466,7 +480,8 @@ class TestTtsNatsAdapter:
                 return_value={"mock": _stub_engine_factory(tmp_path, gate=gate)},
             ):
                 with patch(
-                    "voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+                    "voicecli.adapters.nats.tts_adapter.scoped_path",
+                    side_effect=_patched_scoped_path,
                 ):
                     handle_task = asyncio.create_task(adapter.handle(msg, payload))
                     hb_task = asyncio.create_task(adapter._heartbeat_loop())  # type: ignore[attr-defined]
@@ -571,7 +586,7 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-space", engine="a b")
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available") as mock_avail,
+            patch("voicecli.adapters.nats.tts_adapter._engine_available") as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
@@ -594,7 +609,7 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-wild", engine="*.tts")
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available") as mock_avail,
+            patch("voicecli.adapters.nats.tts_adapter._engine_available") as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
@@ -617,7 +632,7 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-dbl-wild", engine="**")
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available") as mock_avail,
+            patch("voicecli.adapters.nats.tts_adapter._engine_available") as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
@@ -645,7 +660,7 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-ws", engine=bad_engine)
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available") as mock_avail,
+            patch("voicecli.adapters.nats.tts_adapter._engine_available") as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
@@ -667,7 +682,9 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-dotted", engine="engine.with.dots")
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=False) as mock_avail,
+            patch(
+                "voicecli.adapters.nats.tts_adapter._engine_available", return_value=False
+            ) as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
@@ -694,12 +711,16 @@ class TestTtsNatsAdapter:
             return tmp_path / f"{rid}.{ext}"
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True) as mock_avail,
+            patch(
+                "voicecli.adapters.nats.tts_adapter._engine_available", return_value=True
+            ) as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
             ),
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
         ):
             asyncio.run(adapter.handle(msg, payload))
             mock_avail.assert_called_once_with("mock")
@@ -726,12 +747,16 @@ class TestTtsNatsAdapter:
             return tmp_path / f"{rid}.{ext}"
 
         with (
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True) as mock_avail,
+            patch(
+                "voicecli.adapters.nats.tts_adapter._engine_available", return_value=True
+            ) as mock_avail,
             patch(
                 "voicecli.engine._get_registry",
                 return_value={"mock": _stub_engine_factory(tmp_path)},
             ),
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
         ):
             asyncio.run(adapter.handle(msg, payload))
             mock_avail.assert_called_once_with("mock")
@@ -779,8 +804,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -877,8 +904,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -906,8 +935,10 @@ class TestTtsNatsAdapter:
             raise ValueError("unsupported language: zz")
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -938,8 +969,10 @@ class TestTtsNatsAdapter:
             raise ValueError("unsupported language: en")
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -973,8 +1006,10 @@ class TestTtsNatsAdapter:
         payload = _valid_payload(request_id="req-wf")
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -1001,8 +1036,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -1071,8 +1108,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -1113,8 +1152,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
@@ -1150,8 +1191,10 @@ class TestTtsNatsAdapter:
             return None
 
         with (
-            patch("voicecli.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path),
-            patch("voicecli.nats.tts_adapter._engine_available", return_value=True),
+            patch(
+                "voicecli.adapters.nats.tts_adapter.scoped_path", side_effect=_patched_scoped_path
+            ),
+            patch("voicecli.adapters.nats.tts_adapter._engine_available", return_value=True),
             patch("voicecli.api.generate", side_effect=_fake_generate),
         ):
             asyncio.run(adapter.handle(msg, payload))
